Unify printer label
